requirements.txt
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
chromadb>=0.4.22,<0.5.0
llama-index==0.10.65
llama-index-embeddings-huggingface==0.1.4
llama-index-vector-stores-chroma==0.1.3
llama-index-llms-openai>=0.1.27,<0.2.0
openai>=1.12.0,<2.0.0
python-dotenv==1.0.0
pandas>=2.2.0
numpy>=1.26.0
